refactor(desktop): extract network settings helpers
